Post Competition

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/main/Hardware.java
@@ -13,8 +13,8 @@ public class Hardware {
 
     // Prepare variables for encoder use
     // http://www.revrobotics.com/content/docs/Encoder-Guide.pdf
-    private final double HD_HEX_REV_COUNTS = 1120; // 1120 for 40:1, 560 for 20:1
-    private final double DRIVE_GEAR_RATIO = (double)72/(double)125; // This is < 1.0 if geared up
+    private final double HD_HEX_REV_COUNTS = 1120.0; // 1120 for 40:1, 560 for 20:1
+    private final double DRIVE_GEAR_RATIO = 72.0/125.0; // This is < 1.0 if geared up
     private final double WHEEL_CIRCUMFERENCE = 6 * 3.1415;
     public final int DRIVE_COUNTS_PER_INCH = (int) (HD_HEX_REV_COUNTS * DRIVE_GEAR_RATIO *
             WHEEL_CIRCUMFERENCE);
